requirements.txt
apache-airflow-providers-mysql==5.3.0
apache-airflow-providers-apache-spark==4.1.0
pyspark==3.5.0
mysqlclient==2.2.1           # Cliente MySQL para Python
pandas==2.0.3                # Opcional: Si usas Pandas en PySpark